backend/chat_api/requirements.txt
fastapi
uvicorn[standard]
redis
google-genai